Update
Move AdminPage and Controller to SampleController and change admin
