+++ src/main/resources/com/example/stackapp/pages/sample.fxml
@@ -20,6 +20,8 @@
 <?import javafx.scene.text.Font?>
 <?import javafx.scene.text.Text?>
 
+<?import javafx.scene.control.TableView?>
+<?import javafx.scene.control.TableColumn?>
 <BorderPane prefHeight="700.0" prefWidth="900.0" xmlns="http://javafx.com/javafx/18" xmlns:fx="http://javafx.com/fxml/1"
             fx:controller="com.example.stackapp.controller.SampleController">
     <top>
@@ -115,7 +117,7 @@
                     </Button>
                     <Button fx:id="addWorkerBtn" mnemonicParsing="false" onAction="#changePanToAddWorkerPan"
                             prefHeight="38.0" prefWidth="200.0" style="-fx-background-color: #AAB2BD;"
-                            stylesheets="@../colors.css" text="ADD WORKER" visible="true">
+                            stylesheets="@../colors.css" text="ADD WORKER">
                         <font>
                             <Font size="16.0"/>
                         </font>
@@ -790,7 +792,86 @@
                 </children>
             </Pane>
             <Pane fx:id="addWorkerPane" prefHeight="645.0" prefWidth="700.0" style="-fx-background-color: #dce2e8;"
-                  visible="false"/>
+                  visible="false">
+                <Label layoutX="180.0" layoutY="510.0" prefHeight="20.0" prefWidth="84.0" text="USERNAME">
+                    <font>
+                        <Font size="16.0"/>
+                    </font>
+                </Label>
+                <Label layoutX="180.0" layoutY="475.0" prefHeight="20.0" prefWidth="88.0" text="PASSWORD">
+                    <font>
+                        <Font size="16.0"/>
+                    </font>
+                </Label>
+                <Label layoutX="180.0" layoutY="435.0" text="LAST NAME">
+                    <font>
+                        <Font size="16.0"/>
+                    </font>
+                </Label>
+                <Label layoutX="180.0" layoutY="399.0" text="FIRST NAME">
+                    <font>
+                        <Font size="16.0"/>
+                    </font>
+                </Label>
+                <Pane>
+                </Pane>
+                <TableView fx:id="table_users" layoutX="35.0" layoutY="41.0" onMouseClicked="#handleMouseAction"
+                           prefHeight="298.0" prefWidth="613.0">
+                    <columns>
+                        <TableColumn fx:id="tc_id" prefWidth="88.0" text="ID"/>
+                        <TableColumn fx:id="tc_first_name" prefWidth="134.0" text="FIRST NAME"/>
+                        <TableColumn fx:id="tc_last_name" prefWidth="120.0" text="LAST NAME"/>
+                        <TableColumn fx:id="tc_password" prefWidth="155.0" text="PASSWORD"/>
+                        <TableColumn fx:id="tc_username" prefWidth="115.0" text="USERNAME"/>
+                    </columns>
+                </TableView>
+                <VBox layoutX="271.0" layoutY="393.0" prefHeight="200.0" prefWidth="142.0" spacing="10.0">
+                    <children>
+                        <TextField fx:id="tf_first_name">
+                            <font>
+                                <Font size="14.0"/>
+                            </font>
+                        </TextField>
+                        <TextField fx:id="tf_last_name">
+                            <font>
+                                <Font size="14.0"/>
+                            </font>
+                        </TextField>
+                        <TextField fx:id="tf_password" onMouseClicked="#setUsername">
+                            <font>
+                                <Font size="14.0"/>
+                            </font>
+                        </TextField>
+                        <TextField fx:id="tf_username">
+                            <font>
+                                <Font size="14.0"/>
+                            </font>
+                        </TextField>
+                    </children>
+                </VBox>
+                <Button fx:id="btn_insert" layoutX="467.0" layoutY="399.0" mnemonicParsing="false" onMouseClicked="#insert"
+                        prefHeight="30.0" prefWidth="64.0" text="Insert">
+                    <font>
+                        <Font size="14.0"/>
+                    </font>
+                </Button>
+                <Button fx:id="btn_update" layoutX="467.0" layoutY="449.0" mnemonicParsing="false" onAction="#update"
+                        onMouseClicked="#update" prefHeight="30.0" prefWidth="64.0" text="Update">
+                    <font>
+                        <Font size="14.0"/>
+                    </font>
+                </Button>
+                <Button fx:id="btn_delete" layoutX="467.0" layoutY="500.0" mnemonicParsing="false" onMouseClicked="#delete"
+                        prefHeight="30.0" prefWidth="64.0" text="Delete">
+                    <font>
+                        <Font size="14.0"/>
+                    </font>
+                </Button>
+
+
+
+
+            </Pane>
             <Pane fx:id="shelfPane" prefHeight="645.0" prefWidth="700.0" style="-fx-background-color: #dce2e8;"
                   visible="false">
                 <children>
